app/requirements.txt
fastapi==0.109.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.7
uvicorn==0.27.0.post1